Change the ATHENA_KRB4 autoconf macro to not bail if --with-krb4 is passed but krb4 isn't available.
I'm handling this at the autoconf level instead of at the debian/rules
level because autoconf is in a much better position to detect if krb4
is actually present.

discuss and lert are the only two packages that use ATHENA_KRB4
instead of ATHENA_KRB4_REQUIRED, and both of those packages can deal
with only having krb5.