Add minimal DNS helper application
Add a minimal DNS resolver (built for both i386 and amd64) so that we
can stop using the dns-cgi hack.  This minimal DNS resolver uses
libares (c-ares).  It also is capable of validating an IPv4 address,
and returning its output in a format suitable for passing to a Bourne
shell "eval".

Put the necessary targets in the Makefile to build this.

Declare the "host" target as PHONY, and while we're at it, declare
other .PHONY targets that should have already been there ("clean" and
"install")